save 시 password 해싱 하여 로그인 불가 버그 수정 및 edit, delete 로직에서 적절한 권한(게시자)가 아니면 불가능하게 수정

+++ src/controllers/videoController.js
@@ -23,10 +23,13 @@ export const watch = async (req, res) => {
 };
 export const getEdit = async (req, res) => {
     const { id } = req.params;
-    
+    const { user: {_id} } = req.session;
     const videoItem = await Video.findById(id);
     if(!videoItem) {
         return res.status(404).render("404", {pageTitle : "Video not found"});
+    }
+    if(String(videoItem.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     return res.render("edit", {pageTitle : `Edit : ${videoItem.title}`, video: videoItem});
     
@@ -34,10 +37,14 @@ export const getEdit = async (req, res) => {
 
 export const postEdit = async (req, res) => {
     const { id } = req.params;
+    const { user: {_id} } = req.session;
     const { title, description, hashtags } = req.body
     const videoItem = await Video.exists({_id:id});
     if(!videoItem) {
         return res.status(404).render("404", {pageTitle : "Video not found"});
+    }
+    if(String(videoItem.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title,
@@ -81,7 +88,14 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const { id } = req.params
-
+    const { user: {_id} } = req.session;
+    const videoItem = await Video.findById(id);
+    if(!videoItem) {
+        return res.status(404).render("404", {pageTitle : "Video not found"});
+    }
+    if(String(videoItem.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 }
